+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Button component (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
+- Button component tokens for all themes (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Bullet list component tokens for Orange theme (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Split color charts semantic tokens to dedicated provider (Orange-OpenSource/ouds-ios#856)
 - Swift package `swift-docc-plugin` from v1.4.3 to v1.4.5